test(workspaces): fix invites test
